// denops/k8s/models/IoK8sApiCoreV1ContainerStatus.ts
/**
 * Kubernetes
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1.24.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { IoK8sApiCoreV1ContainerState } from './IoK8sApiCoreV1ContainerState.ts';
import { HttpFile } from '../http/http.ts';

/**
* ContainerStatus contains details for the current status of this container.
*/
export class IoK8sApiCoreV1ContainerStatus {
    /**
    * Container's ID in the format '<type>://<container_id>'.
    */
    'containerID'?: string;
    /**
    * The image the container is running. More info: https://kubernetes.io/docs/concepts/containers/images.
    */
    'image': string;
    /**
    * ImageID of the container's image.
    */
    'imageID': string;
    'lastState'?: IoK8sApiCoreV1ContainerState;
    /**
    * This must be a DNS_LABEL. Each container in a pod must have a unique name. Cannot be updated.
    */
    'name': string;
    /**
    * Specifies whether the container has passed its readiness probe.
    */
    'ready': boolean;
    /**
    * The number of times the container has been restarted.
    */
    'restartCount': number;
    /**
    * Specifies whether the container has passed its startup probe. Initialized as false, becomes true after startupProbe is considered successful. Resets to false when the container is restarted, or if kubelet loses state temporarily. Is always true when no startupProbe is defined.
    */
    'started'?: boolean;
    'state'?: IoK8sApiCoreV1ContainerState;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "containerID",
            "baseName": "containerID",
            "type": "string",
            "format": ""
        },
        {
            "name": "image",
            "baseName": "image",
            "type": "string",
            "format": ""
        },
        {
            "name": "imageID",
            "baseName": "imageID",
            "type": "string",
            "format": ""
        },
        {
            "name": "lastState",
            "baseName": "lastState",
            "type": "IoK8sApiCoreV1ContainerState",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "ready",
            "baseName": "ready",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "restartCount",
            "baseName": "restartCount",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "started",
            "baseName": "started",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "state",
            "baseName": "state",
            "type": "IoK8sApiCoreV1ContainerState",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return IoK8sApiCoreV1ContainerStatus.attributeTypeMap;
    }

    public constructor() {
    }
}